app: tests: add small test framework
Initial test and unittest framework. It will run test on the target
device

Since it will make the final binary size increase by a big amount, the
test framework is disabled by default and it should only be enabled on
targets with no memory restrictions.

This can be used for an emulator target in order to validate the library
and it can be part of the validation done before merging new changes

Signed-off-by: Fernando Lugo <fernando@gmail.com>