lib: fmt: remove dependency on board peripherals
After adding support for more peripherals it is easy to create circular
dependencies, since fmt library is used everything for printing and
since board peripherals depends on multiple drivers it is very simple to
create a circular dependency. That's why fmt should only have
dependencies on simple things, such as string or device::console
interface.

Instead create an API to register the console device that will be used
as standard output that can be called as soon as the console is ready.
That will also make it easy to add support for multiple output consoles
in the future

Signed-off-by: Fernando Lugo <fernando@gmail.com>