create dependencies when compiling files
This can handle C and C++ dependencies when rebuilding. First build
dependencies do not exist but that is not an issue for normal C and C++
files. However, for modules it is different, since some modules depend
on other modules, so in order to generate the pcm file it needs the pcm
files from the dependencies. So module dependencies needs to be handled
in a different way.

Signed-off-by: Fernando Lugo <fernando@gmail.com>